Move sheet to a separate component
